requirements.txt
streamlit==1.40.0
langchain==0.1.20
langchain-openai==0.1.8
langchain-core==0.1.50
openai==1.42.0
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
openpyxl==3.1.5
requests==2.31.0
beautifulsoup4==4.12.0
